requirements.txt
einops
torchdata